New SFX + SFX Fixes
Added enemy squish and stem borer shooting sfx.

Fixed issues with certain sfx not playing due to merges and such.

Have still left out my attempts to add footsteps at this point, as its taking a while to get right.

Known issue I want to fix is the checkpoint activate sfx replaying when walking over already lit checkpoints. This is due to the checkpoint still attempting to update the players spawn location even when it is already the active checkpoint